Add `synced` event to matrix client
This allows to wait for the next sync to make sure that you're working
on sufficiently up-to-date data.

This is required to make the PFS wait for the first sync until it
calculated routes. Otherwise it won't have valid presence information
for a while. But it's probably also useful in other cases.