Simplify by avoiding a greenlet
It is not important to call `_handle_response` in a greenlet. We wait
for it to finish anyway before calling it the next time.

However, this will increase the latency a bit because we only do the
next sync after the handling has been done. The different should not be
noticeable in most cases, but it's hard to tell for sure.