Batching message processing
Changes the way Matrix messages are processed, instead of dispatch each
mesage individually, this batches all messages from the sync together,
and then dispatch them all in a single transaction.

The benefit main benefit of this approach is the reduce number of
deepcopies, which can have a signification implication in performance.