Fixes bad `check_gas_reserve` performance.
The task `check_gas_reserve` was creating a token network proxy for
every existing token network. This should be avoided because the proxy
constructor will check the target address`s bytecode, to detect
configuration errors. The problem is that for a large registry, e.g.
>=100 token networks, executing the aforementioned test will take a
while. This is specially bad since the tokens are only necessary to get
the counter of current pending channel opennings, which is 0 for new
token network proxy (If the proxy does not exist then there is no
pending transactoins for it).

This commit changes the code to, instead of looping over all token
network addresses, to loop over the currently instantiated proxies. This
means the cost of instantiating the proxies will be amortized and only
paid when necessary.

Note: This problem will go away after #1621 is implemented.