Use time between syncs for timeout
We need to call sync every 30s or we'll be considered unreachable. The
old check does not catch all of those cases. Example:

0s: Sync started
20s: Sync returned
40s: Next sync started

The old check would check 40s - 20s <= 20s and succeed, but matrix will
see the node as offline for a while because the time between syncs is
greater than 30s.
The new check will notice that 40s - 0s > 22s and throws an exception.

Relates to https://github.com/raiden-network/raiden/issues/5635.