Fix Matrix server sorting
In #5544 the algorithm to sort the available Matrix servers by response
time was changed to only consider the first result.

As described in #5626, this caused some issues, namely:
- If a request failed (e.g. name resolution, network unreachable) that
  greenlet would return first, causing a false positive.
- The `make_client()` function that is using `sort_servers_closest()`
  has the option of falling back to a less optimal server in case the
  'best' one (by response time) turned out to be unusable for some other
  reason. This was no longer possible with only one result.

This changes the algorithm back to return multiple results but still limit
the worst case time we wait to gather those to a reasonable amount
(7s currently), so that slow servers can't arbitrarily delay node startup,
which was the initial motivation for #5544.