Kill long running threads
Code to support improving performance of the client. This will kill the
blocking thread, if any, which will result in the death of the node and
the stack trace of the slow thread printed.

Notes:
- The monitoring thread only gets a chance to run once a
context-switch happens, meaning the stack traces will be produced in the
points were IO is done.
- This does not consider the time used to run a single task, but the
  continuos amount of time used by a single thread without yielding to
  the event loop.