Batch processing of the current queue
The previous code was calling _handle_response individually for each
entry in the queue. Which means each response would be a separate set of
StateChanges. With this change all the currenlty pending messages are
processed in a single batch, this reduces the number of deepcopies, and
databases writes from `n` to `1`.