bugfix: healthcheck and whitelist is not the same thing
An address that is whitelist does not necessarily have the healthcheck
running, this means that checking if the address is available in the
address manager is not sufficient, one also has to check the current
value for the availability information. If the node state is anything
but UNKNWON we know the presence information was used at least once (it
does not, however, mean the presence information is up-to-date).