Check sync timeout during all iterations
The exemption has been introduced when we counted the time between sync
starts. This was a problem for initial (potentially slow) syncs. Now we
check the time between syncs, which will only fail when the raiden node
is blocked by too high load or a bug. We do want an error in those
cases, so the exemption does not make sense anymore and only masks the
problem which will manifest as nodes being treated as offline.

Closes https://github.com/raiden-network/raiden/issues/5898