fix nursery subprocess registration race condition
The previous version of the load script had a race condition were the
grenlet for `run` executed after the Janitor context manager exited.
This allowed `Popen` to execute and start the child process, but the
registration code didn't run (leading the child process orphan).

To fix the above race, a lock was introduce. The lock protects the
`stop_event`, so that if the `exec_under_watch` know the process is not
shuting down while it is executed, which allows it to safely start a
subprocess and registered it.