merged `allocate_next_slot` into `transact`.
The method `allocate_next_slot` existed for 3 reasons:

- Allow concurrent transaction to be instantiated (without waiting for
  the precondition/postcondition validation).
- Protect the `_available_nonce` with a lock, since the `nonce` must be
  used only once (ignoring transaction overwrites). The lock exists to
  ensure only one transaction will use the given nonce.
- Wrap the value of the `nonce` in a poors man linear type, to log and
  throw an exception when the allocated nonce is not used to avoid
  stalling the account's transaction queue.

However, the code had an annoying corner case. If a transaction is
created using values that cannot be correct serialized, or if the
blockchain node rejects the transaction, the nonce would be allocated
for a transaction that would never be sent. If another transaction is
allocated just after, to use the next nonce, which happens to be a
valid transaction, a hole into the transaction sequence would be
introduced leading to the stagnation of the transaction queue.

This removes the method `allocate_next_slot` since it is only used by
`transact`. But preserves the slot abstraction to make sure a log is
printed if any of the above happens. At the same time the critical
sectin of the `_nonce_lock` increase slightly to include the
`eth_sendRawTransaction` REST call, the nonce is only increment
afterwards. This should allow Raiden to gracefully recover from the
above, since the nonce is only incremented locally if the transaction
was effectively sent and accepted by the ethereum node.