Added UserDeposit.deposit
In contrast to the `approve_and_deposit`, this method assumes the
allowance was already set. This is a nice optimization for the scenario
player, were the orchestration account is used to fund an unbounded number
of raiden accounts. This effectively reduces the number of transactions
in half.

There are a few downsides to this approve:

- It can not guarantee the transaction will be successful, because
  the allowance may be too low.
- It can be error prone, since a call to `approve_and_deposit` will
  overwrite the existing allowance, making for a hard to debug corner
  cases.