Print the exception to stderr
The previous approach redirectered the stacktrace to a file to avoid
showing it to a end users, however it makes it harder for our tooling to
make the stacktrace available for debugging, since the scenarios are
executed inside a docker container and the tmp folder is not mapped.
This changes the code to print the exception to the stderr to make it
available for debugging.